+++ internal/service/golden_gate/golden_gate_deployment_peers_data_source.go
@@ -1,0 +1,235 @@
+// Copyright (c) 2017, 2024, Oracle and/or its affiliates. All rights reserved.
+// Licensed under the Mozilla Public License v2.0
+
+package golden_gate
+
+import (
+	"context"
+
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+	oci_golden_gate "github.com/oracle/oci-go-sdk/v65/goldengate"
+
+	"github.com/oracle/terraform-provider-oci/internal/client"
+	"github.com/oracle/terraform-provider-oci/internal/tfresource"
+)
+
+func GoldenGateDeploymentPeersDataSource() *schema.Resource {
+	return &schema.Resource{
+		Read: readGoldenGateDeploymentPeers,
+		Schema: map[string]*schema.Schema{
+			"filter": tfresource.DataSourceFiltersSchema(),
+			"deployment_id": {
+				Type:     schema.TypeString,
+				Required: true,
+			},
+			"display_name": {
+				Type:     schema.TypeString,
+				Optional: true,
+			},
+			"state": {
+				Type:     schema.TypeString,
+				Optional: true,
+			},
+			"deployment_peer_collection": {
+				Type:     schema.TypeList,
+				Computed: true,
+				Elem: &schema.Resource{
+					Schema: map[string]*schema.Schema{
+						// Required
+
+						// Optional
+
+						// Computed
+						"items": {
+							Type:     schema.TypeList,
+							Computed: true,
+							Elem: &schema.Resource{
+								Schema: map[string]*schema.Schema{
+									// Required
+
+									// Optional
+
+									// Computed
+									"availability_domain": {
+										Type:     schema.TypeString,
+										Computed: true,
+									},
+									"deployment_id": {
+										Type:     schema.TypeString,
+										Computed: true,
+									},
+									"display_name": {
+										Type:     schema.TypeString,
+										Computed: true,
+									},
+									"fault_domain": {
+										Type:     schema.TypeString,
+										Computed: true,
+									},
+									"peer_role": {
+										Type:     schema.TypeString,
+										Computed: true,
+									},
+									"peer_type": {
+										Type:     schema.TypeString,
+										Computed: true,
+									},
+									"region": {
+										Type:     schema.TypeString,
+										Computed: true,
+									},
+									"state": {
+										Type:     schema.TypeString,
+										Computed: true,
+									},
+									"time_created": {
+										Type:     schema.TypeString,
+										Computed: true,
+									},
+									"time_role_changed": {
+										Type:     schema.TypeString,
+										Computed: true,
+									},
+									"time_updated": {
+										Type:     schema.TypeString,
+										Computed: true,
+									},
+								},
+							},
+						},
+					},
+				},
+			},
+		},
+	}
+}
+
+func readGoldenGateDeploymentPeers(d *schema.ResourceData, m interface{}) error {
+	sync := &GoldenGateDeploymentPeersDataSourceCrud{}
+	sync.D = d
+	sync.Client = m.(*client.OracleClients).GoldenGateClient()
+
+	return tfresource.ReadResource(sync)
+}
+
+type GoldenGateDeploymentPeersDataSourceCrud struct {
+	D      *schema.ResourceData
+	Client *oci_golden_gate.GoldenGateClient
+	Res    *oci_golden_gate.ListDeploymentPeersResponse
+}
+
+func (s *GoldenGateDeploymentPeersDataSourceCrud) VoidState() {
+	s.D.SetId("")
+}
+
+func (s *GoldenGateDeploymentPeersDataSourceCrud) Get() error {
+	request := oci_golden_gate.ListDeploymentPeersRequest{}
+
+	if deploymentId, ok := s.D.GetOkExists("deployment_id"); ok {
+		tmp := deploymentId.(string)
+		request.DeploymentId = &tmp
+	}
+
+	if displayName, ok := s.D.GetOkExists("display_name"); ok {
+		tmp := displayName.(string)
+		request.DisplayName = &tmp
+	}
+
+	if state, ok := s.D.GetOkExists("state"); ok {
+		request.LifecycleState = oci_golden_gate.ListDeploymentPeersLifecycleStateEnum(state.(string))
+	}
+
+	request.RequestMetadata.RetryPolicy = tfresource.GetRetryPolicy(false, "golden_gate")
+
+	response, err := s.Client.ListDeploymentPeers(context.Background(), request)
+	if err != nil {
+		return err
+	}
+
+	s.Res = &response
+	request.Page = s.Res.OpcNextPage
+
+	for request.Page != nil {
+		listResponse, err := s.Client.ListDeploymentPeers(context.Background(), request)
+		if err != nil {
+			return err
+		}
+
+		s.Res.Items = append(s.Res.Items, listResponse.Items...)
+		request.Page = listResponse.OpcNextPage
+	}
+
+	return nil
+}
+
+func (s *GoldenGateDeploymentPeersDataSourceCrud) SetData() error {
+	if s.Res == nil {
+		return nil
+	}
+
+	s.D.SetId(tfresource.GenerateDataSourceHashID("GoldenGateDeploymentPeersDataSource-", GoldenGateDeploymentPeersDataSource(), s.D))
+	resources := []map[string]interface{}{}
+	deploymentPeer := map[string]interface{}{}
+
+	items := []interface{}{}
+	for _, item := range s.Res.Items {
+		items = append(items, DeploymentPeerSummaryToMap(item))
+	}
+	deploymentPeer["items"] = items
+
+	if f, fOk := s.D.GetOkExists("filter"); fOk {
+		items = tfresource.ApplyFiltersInCollection(f.(*schema.Set), items, GoldenGateDeploymentPeersDataSource().Schema["deployment_peer_collection"].Elem.(*schema.Resource).Schema)
+		deploymentPeer["items"] = items
+	}
+
+	resources = append(resources, deploymentPeer)
+	if err := s.D.Set("deployment_peer_collection", resources); err != nil {
+		return err
+	}
+
+	return nil
+}
+
+func DeploymentPeerSummaryToMap(obj oci_golden_gate.DeploymentPeerSummary) map[string]interface{} {
+	result := map[string]interface{}{}
+
+	if obj.AvailabilityDomain != nil {
+		result["availability_domain"] = string(*obj.AvailabilityDomain)
+	}
+
+	if obj.DeploymentId != nil {
+		result["deployment_id"] = string(*obj.DeploymentId)
+	}
+
+	if obj.DisplayName != nil {
+		result["display_name"] = string(*obj.DisplayName)
+	}
+
+	if obj.FaultDomain != nil {
+		result["fault_domain"] = string(*obj.FaultDomain)
+	}
+
+	result["peer_role"] = string(obj.PeerRole)
+
+	result["peer_type"] = string(obj.PeerType)
+
+	if obj.Region != nil {
+		result["region"] = string(*obj.Region)
+	}
+
+	result["state"] = string(obj.LifecycleState)
+
+	if obj.TimeCreated != nil {
+		result["time_created"] = obj.TimeCreated.String()
+	}
+
+	if obj.TimeRoleChanged != nil {
+		result["time_role_changed"] = obj.TimeRoleChanged.String()
+	}
+
+	if obj.TimeUpdated != nil {
+		result["time_updated"] = obj.TimeUpdated.String()
+	}
+
+	return result
+}