+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 6.35.0 (April 16, 2025)
+
+### Added
+- Support for Add Cross-AD DR for GGS deployment in a multi-AD region
+- Security Attributes feature to Load Balancer Service
+### Bug Fix
+- Container Instance HTTP health check port field set to required
+- Fix for privateenpointip and displayName params in update
+- For for zpr provider updates
+
 ## 6.34.0 (April 09, 2025)
 
 ### Added